+++ src/components/molecules/Footers/RoomFooter.tsx
@@ -9,7 +9,6 @@ interface Props {
 export default ({ className }: Props) => (
   <div
     css={css`
-      background-color: #00ffff;
       height: 100px;
     `}
     className={className}
