+++ src/components/organisms/Room/RightColumn.tsx
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 
-import { jsx, css } from '@emotion/core';
+import { jsx } from '@emotion/core';
 
 interface Props {
   className?: string;
@@ -9,12 +9,7 @@ interface Props {
 }
 
 export default ({ className, isEditing, setIsEditing }: Props) => (
-  <div
-    css={css`
-      background-color: #00ff00;
-    `}
-    className={className}
-  >
+  <div className={className}>
     これはTabです
     <button type="button" onClick={() => setIsEditing(!isEditing)}>
       {isEditing ? '編集を終了する' : '編集する'}
