+++ src/components/organisms/VideoSearch/index.tsx
@@ -9,7 +9,6 @@ interface Props {
 export default ({ className }: Props) => (
   <div
     css={css`
-      background-color: #0000ff;
       height: 200vh;
     `}
     className={className}
